requirements.txt
pymilvus==2.4.4
python-dotenv==1.0.1
sentence-transformers==3.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
pydantic==2.7.1
pydantic_core==2.18.2
uvicorn==0.29.0
ibm_watson_machine_learning==1.0.356